chore: reorganize folders
