+++ django_forest/authentication/views/callback.py
@@ -10,7 +10,7 @@ from django_forest.authentication.oidc.client_manager import OidcClientManager
 from django_forest.authentication.utils import get_callback_url
 from django_forest.utils.error_handler import MESSAGES
 from django_forest.utils.forest_api_requester import ForestApiRequester
-from django_forest.utils.get_forest_setting import get_forest_setting
+from django_forest.utils.forest_setting import get_forest_setting
 
 
 logger = logging.getLogger(__name__)
