+++ django_forest/resources/associations/utils/__init__.py
@@ -1,0 +1,3 @@
+from django_forest.resources.associations.utils.association import AssociationMixin
+
+__all__ = ['AssociationMixin']
