+++ django_forest/resources/views/detail.py
@@ -2,7 +2,7 @@ from django.http import JsonResponse, HttpResponse
 from django.views import generic
 
 from django_forest.resources.utils import SmartFieldMixin, FormatFieldMixin, JsonApiSerializerMixin, ResourceMixin
-from django_forest.utils.json_api_serializer import JsonApiSchema
+from django_forest.utils.schema.json_api_schema import JsonApiSchema
 from django_forest.utils.models import Models
