+++ django_forest/utils/schema/__init__.py
@@ -10,8 +10,8 @@ from django.utils.module_loading import autodiscover_modules
 
 from django_forest.utils.schema.apimap_errors import APIMAP_ERRORS
 from django_forest.utils.models import Models
-from django_forest.utils.get_type import get_type
-from django_forest.utils.json_api_serializer import create_json_api_schema
+from django_forest.utils.type_mapping import get_type
+from django_forest.utils.schema.json_api_schema import create_json_api_schema
 from django_forest.utils.forest_api_requester import ForestApiRequester
 from .definitions import COLLECTION, FIELD
 from .validations import handle_validations
@@ -21,7 +21,7 @@ from .default import handle_default_value
 from .version import get_app_version
 
 from .. import get_accessor_name
-from ..get_forest_setting import get_forest_setting
+from ..forest_setting import get_forest_setting
 
 # Get an instance of a logger
 logger = logging.getLogger(__name__)
