+++ django_forest/utils/type_mapping.py
